# Conflicts:
#	android/AndroidManifest.xml
#	android/src/com/mapswithme/maps/MwmActivity.java
#	android/src/com/mapswithme/maps/SplashActivity.java
#	android/src/com/mapswithme/maps/widget/placepage/PlacePageView.java
#	tools/unix/version.sh

+++ android/jni/com/mapswithme/maps/SearchEngine.cpp
@@ -40,7 +40,7 @@ FeatureID const kEmptyFeatureId;
 Results g_results;
 
 // Timestamp of last search query. Results with older stamps are ignored.
-uint64_t g_queryTimestamp;
+jlong g_queryTimestamp;
 // Implements 'NativeSearchListener' java interface.
 jobject g_javaListener;
 jmethodID g_updateResultsId;
@@ -75,7 +75,7 @@ jobject ToJavaResult(Result & result, search::ProductInfo const & productInfo, b
   jni::TScopedLocalIntArrayRef ranges(
       env, env->NewIntArray(static_cast<jsize>(result.GetHighlightRangesCount() * 2)));
   jint * rawArr = env->GetIntArrayElements(ranges, nullptr);
-  for (int i = 0; i < result.GetHighlightRangesCount(); i++)
+  for (size_t i = 0; i < result.GetHighlightRangesCount(); i++)
   {
     auto const & range = result.GetHighlightRange(i);
     rawArr[2 * i] = range.first;
@@ -147,7 +147,7 @@ jobject ToJavaResult(Result & result, search::ProductInfo const & productInfo, b
 }
 
 void OnResults(Results const & results, vector<search::ProductInfo> const & productInfo,
-               long long timestamp, bool isMapAndTable, bool hasPosition, double lat, double lon)
+               jlong timestamp, bool isMapAndTable, bool hasPosition, double lat, double lon)
 {
   // Ignore results from obsolete searches.
   if (g_queryTimestamp > timestamp)
@@ -159,13 +159,12 @@ void OnResults(Results const & results, vector<search::ProductInfo> const & prod
   {
     jni::TScopedLocalObjectArrayRef jResults(
         env, BuildSearchResults(results, productInfo, hasPosition, lat, lon));
-    env->CallVoidMethod(g_javaListener, g_updateResultsId, jResults.get(),
-                        static_cast<jlong>(timestamp));
+    env->CallVoidMethod(g_javaListener, g_updateResultsId, jResults.get(), timestamp);
   }
 
   if (results.IsEndMarker())
   {
-    env->CallVoidMethod(g_javaListener, g_endResultsId, static_cast<jlong>(timestamp));
+    env->CallVoidMethod(g_javaListener, g_endResultsId, timestamp);
     if (isMapAndTable && results.IsEndedNormal())
       g_framework->NativeFramework()->GetSearchAPI().PokeSearchInViewport();
   }
